Implemented NavBar and Footer grouped functionality implemented all NavBar buttons implemented all Footer links populated pages with components
