+++ frontend/src/pages/about.tsx
@@ -2,6 +2,8 @@ import React from "react";
 import "../app/about.css";
 import MapComponent from "@/components/organisms/Map";
 import TopPanel from "@/components/organisms/TopPanel";
+import NavBar from "../components/molecules/NavBar"; // Adjusted the path to match the relative structure
+import Footer from "@/components/molecules/Footer"; // Adjusted the path to match your project structure
 // import SimpleMap from "@/components/MapWrapper";
 /** An About page */
 /*const About = () => {
@@ -52,6 +54,8 @@ const teamMembers = [
 
 const About = () => {
   return (
+    <div>
+      <NavBar />
     <div className="container">
       <section className="hero">
         <div className="hero-text">
@@ -118,8 +122,9 @@ const About = () => {
           ))}
       </section>
     </div>
+    <Footer />
+    </div>
   );
 };
 
 export default About;
-//h3y
