+++ frontend/src/pages/landing.tsx
@@ -1,124 +1,17 @@
 import React from 'react';
 import NavBar from "../components/molecules/NavBar";
-import { AppBar, Toolbar, Button, Typography, Box, useTheme, useMediaQuery, Paper, MenuItem, Select, Link} from '@mui/material';
+import {Button, Typography, Box, useTheme, useMediaQuery, Paper, Link} from '@mui/material';
 import { useRouter } from 'next/router'; 
+import Footer from '@/components/molecules/Footer';
 
 const Landing : React.FC = () => {
   const theme = useTheme();
   const router = useRouter();
   const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
-  const [selectedLanguage, setSelectedLanguage] = useState<'en' | 'mn'>('en');
-  const handleLanguageChange = (event: any) => {
-    setSelectedLanguage(event.target.value);
-  };
-  const handleNavigate = (path: string) => {
-    router.push(path);
-  };
 
   return (
     <div>
-      <AppBar
-      position="static"
-      color="transparent"
-      elevation={0}
-      sx={{ backgroundColor: '#E6EEEC' }} 
-      >
-      <Toolbar sx={{ justifyContent: 'space-between'}}>
-        <Typography variant="h6" component="div" sx={{
-          color: '#065143', fontWeight: 'bold',
-          fontSize: isMobile ? '24px' : '30px',
-        }}>
-          GreenZone Analytics
-        </Typography>
-        <Box
-          sx={{
-            display: 'flex',
-            flexDirection: isMobile ? 'column' : 'row',
-            gap: 2,
-            backgroundColor: '#E6EEEC',
-            padding: 1.5,
-            borderRadius: 2,
-            marginTop: isMobile ? 2 : 0,
-            alignItems: 'center',
-          }}
-        >
-        <Button variant="text" color="inherit"sx={{ textTransform: 'none', fontSize: '18px',}}>
-            About
-        </Button>
-        <Button variant="text" color="inherit"sx={{ textTransform: 'none', fontSize: '18px', }} onClick={() => handleNavigate('/summary-stats')}>
-          Insights
-        </Button>
-        <Button variant="text" color="inherit"sx={{ textTransform: 'none', fontSize: '18px',}}>
-          Methodologies
-        </Button>
-        <Button
-          variant="contained"
-          sx={{
-            bgcolor: '#065143',
-            borderRadius: '12px',
-            '&:hover': { bgcolor: '#043F33' },
-            color: 'white',
-            fontSize: '18px',
-            textTransform: 'none',
-            display: 'flex',
-            alignItems: 'center',
-            gap: 1, 
-            paddingX: 2,
-            paddingY: 1,
-          }} onClick = {() => handleNavigate('/monitoring-platform')}
-        >
-          Launch
-        <img
-          src="/launch_icon.png"
-          alt="icon"
-          style={{ width: 20, height: 20 }}
-        /> 
-      </Button>
-      <Box sx={{ minWidth: 160 }}>
-            <Select
-              value={selectedLanguage}
-              onChange={handleLanguageChange}
-              variant="standard"
-              disableUnderline
-              sx={{
-                backgroundColor: '#E6EEEC',
-                borderRadius: '12px',
-                fontFamily: 'Poppins, sans-serif',
-                fontSize: '16px',
-                paddingY: '4px',
-                paddingX: '8px',
-                '& .MuiSelect-icon': {
-                  right: 10,
-                },
-              }}
-              renderValue={() => (
-                <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                  <img
-                    src={`/flags_${selectedLanguage === 'en' ? 'en' : 'mn'}.png`}
-                    width={20}
-                    alt="flag"
-                  />
-                  <span>{selectedLanguage === 'en' ? 'EN' : 'MN'}</span>
-                </Box>
-              )}
-            >
-              <MenuItem value="en">
-                <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                  <img src="/flags_en.png" width={20} alt="EN" />
-                  EN
-                </Box>
-              </MenuItem>
-              <MenuItem value="mn">
-                <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                  <img src="/flags_mn.png" width={20} alt="MN" />
-                  MN
-                </Box>
-              </MenuItem>
-            </Select>
-          </Box>
-        </Box>
-      </Toolbar>
-    </AppBar>
+    <NavBar />
       <div className="w-full flex flex-col gap-[300px]">
       {/* Hero Section */}
     <section className="w-[1600px] mx-auto flex flex-col gap-[140px]">
@@ -411,69 +304,8 @@ const Landing : React.FC = () => {
       </Box>
       </Box>
     </section>
-
-    {/*Footer Section*/}
-    <Box component="footer" sx={{ backgroundColor: '#4E4E4E', color: 'white', py: 6 }}>
-      <Box
-        sx={{
-          maxWidth: 'lg',
-          mx: 'auto',
-          px: 2,
-          display: 'flex',
-          flexDirection: { xs: 'column', md: 'row' },
-          gap: 4,
-        }}
-      >
-        {/* Branding */}
-        <Box sx={{ flex: 1 }}>
-          <Typography variant="h6" sx={{ mb: 2 }}>
-            GreenZone Analytics
-          </Typography>
-          <Typography variant="body2" sx={{ color: 'white' }}>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit sed do eiusmod tempor.
-          </Typography>
-        </Box>
-
-        {/* Useful Links */}
-        <Box sx={{ flex: 1 }}>
-          <Typography variant="h6" sx={{ mb: 2 }}>
-            Useful Links
-          </Typography>
-            <Box component="ul" sx={{ listStyle: 'none', p: 0, m: 0,}}>
-              {[
-                { href: '/', label: 'Home' },
-                { href: '/about', label: 'About' },
-                { href: '/insights', label: 'Insights' },
-                { href: '/methodologies', label: 'Methodologies' },
-                { href: '/platform', label: 'Launch platform' },
-              ].map((item) => (
-                <Box key={item.href} component="li" sx={{ mb: 1 }}>
-                  <Link
-                    href={item.href}
-                    underline="hover"
-                    sx={{ color: 'white', '&:hover': { color: 'grey-400' } }}
-                  >
-                    {item.label}
-                  </Link>
-                </Box>
-              ))}
-            </Box>
-        </Box>
-
-        {/* Contact */}
-        <Box sx={{ flex: 1 }}>
-          <Typography variant="h6" sx={{ mb: 2 }}>
-            Contact Us
-          </Typography>
-          <Box component="ul" sx={{ listStyle: 'none', p: 0, m: 0, color: 'white' }}>
-            <Box component="li" sx={{ mb: 1 }}>email@organization.com</Box>
-            <Box component="li" sx={{ mb: 1 }}>123-456-7890</Box>
-            <Box component="li">Location, State, 12345</Box>
-          </Box>
-        </Box>
-      </Box>
-    </Box>
-      </div>
+    <Footer />
+    </div>
     </div>
   );
 };
