+++ frontend/src/pages/methodology.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
 import NavBar from "../components/molecules/NavBar";
+import Footer from "../components/molecules/Footer";
 
 const Methodology = () => {
   return (
-    <div className="text-center p-4">
-        <NavBar />
-        <h1 className="text-2xl font-semibold">Hi! This is the Methodology page</h1>
+    <div>
+    <NavBar/>
+    <Footer />
     </div>
+
   );
 };
